src/requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
alabaster==0.7.12
apt-clone==0.2.1
apturl==0.5.2
astor==0.7.1
astroid==2.2.5
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
batchgenerators==0.19.3
beautifulsoup4==4.6.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2019.6.16
chardet==3.0.4
click==6.7
cloudpickle==1.6.0
colorama==0.3.7
colorlover==0.3.0
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.12
dask==1.1.1
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
entrypoints==0.3
Flask==1.1.1
future==0.18.2
gast==0.2.2
google-pasta==0.1.5
googledrivedownloader==0.4
graphviz==0.11.1
grpcio==1.20.0
gym==0.17.3
h5py==2.10.0
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
imagesize==1.1.0
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.17.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kazam==1.5.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
language-selector==0.1
lazy-object-proxy==1.4.1
libarchive==0.4.7
lightgbm==2.3.1
linecache2==1.0.0
llvmlite==0.29.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
mandelbrot==0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
MedPy==0.4.0
mistune==0.8.4
mock==2.0.0
mtcnn==0.0.8
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.3
nose==1.3.7
notebook==6.0.0
numba==0.44.1
numpy==1.17.0
nvidia-ml-py3==7.352.0
onboard==1.4.1
opencv-python==4.0.0.21
opencv-python3==1.0
packaging==20.4
PAM==0.4.2
pandas==1.1.4
pandocfilters==1.4.2
parso==0.5.1
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==3.10.0
portalocker==1.5.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyglet==1.5.0
Pygments==2.4.2
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==2.3.1
pymacaroons==0.13.0
pymp-pypi==0.4.3
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.15.3
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.1
python-debian==0.1.32
python-telegram-bot==10.1.0
python-xapp==1.2.0
python-xlib==0.20
pytz==2020.4
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.3.1
pyzmq==18.0.2
qtconsole==4.5.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
retrying==1.3.3
scikit-build==0.11.1
scikit-image==0.15.0
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.9.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
SimpleITK==1.2.0
six==1.15.0
slackclient==2.1.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
stable-baselines==2.10.1
starlette==0.12.10
system-service==0.3
systemd-python==234
tb-nightly==1.14.0a20190424
tensorboard==1.12.2
tensorboardX==1.8
tensorflow==1.12.0
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019042301
tf-nightly==1.14.1.dev20190424
toolz==0.9.0
torchfile==0.1.0
torchvision-nightly==0.2.1
tornado==6.0.3
traceback2==1.4.0
traitlets==4.3.2
trixi==0.1.2.1
typed-ast==1.4.0
typing==3.7.4.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
umap-learn==0.3.9
unittest2==1.1.0
urllib3==1.25.3
virtualenv==15.1.0
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
widgetsnbextension==3.5.0
wrapt==1.11.2
xkit==0.0.0
yarl==1.3.0
